+++ pybop/optimisers/base_pints_optimiser.py
@@ -10,7 +10,7 @@ from pints import RectangularBoundaries as PintsRectangularBoundaries
 from pints import SequentialEvaluator as PintsSequentialEvaluator
 from pints import strfloat as PintsStrFloat
 
-from pybop import BaseOptimiser, _CuckooSearch
+from pybop import BaseOptimiser
 
 
 class BasePintsOptimiser(BaseOptimiser):
@@ -141,14 +141,14 @@ class BasePintsOptimiser(BaseOptimiser):
                 print(f"NOTE: Boundaries ignored by {self.pints_optimiser}")
                 self.bounds = None
             else:
-                if issubclass(self.pints_optimiser, (PintsPSO, _CuckooSearch)):
+                if issubclass(self.pints_optimiser, PintsPSO):
                     if not all(
                         np.isfinite(value)
                         for sublist in self.bounds.values()
                         for value in sublist
                     ):
                         raise ValueError(
-                            "Either all bounds or no bounds must be set for {self.pints_optimiser.__name__}."
+                            f"Either all bounds or no bounds must be set for {self.pints_optimiser.__name__}."
                         )
                 self._boundaries = PintsRectangularBoundaries(
                     self.bounds["lower"], self.bounds["upper"]